+++ packages/npm/@amazeelabs/publisher/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.30](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/publisher@2.4.29...@amazeelabs/publisher@2.4.30) (2024-05-27)
+
+**Note:** Version bump only for package @amazeelabs/publisher
+
+
+
+
+
 ## [2.4.29](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/publisher@2.4.28...@amazeelabs/publisher@2.4.29) (2024-05-10)
 
 **Note:** Version bump only for package @amazeelabs/publisher