+++ packages/npm/@amazeelabs/token-auth-middleware/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.7](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/token-auth-middleware@1.1.6...@amazeelabs/token-auth-middleware@1.1.7) (2024-05-27)
+
+**Note:** Version bump only for package @amazeelabs/token-auth-middleware
+
+
+
+
+
 ## [1.1.6](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/token-auth-middleware@1.1.5...@amazeelabs/token-auth-middleware@1.1.6) (2024-05-10)
 
 **Note:** Version bump only for package @amazeelabs/token-auth-middleware