+++ packages/tests/silverback-drupal/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.130](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-drupal-tests@1.1.129...@-amazeelabs/silverback-drupal-tests@1.1.130) (2024-06-07)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-drupal-tests
+
+
+
+
+
+## [1.1.129](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-drupal-tests@1.1.128...@-amazeelabs/silverback-drupal-tests@1.1.129) (2024-05-29)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-drupal-tests
+
+
+
+
+
+## [1.1.128](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-drupal-tests@1.1.127...@-amazeelabs/silverback-drupal-tests@1.1.128) (2024-05-27)
+
+**Note:** Version bump only for package @-amazeelabs/silverback-drupal-tests
+
+
+
+
+
 ## [1.1.127](https://github.com/AmazeeLabs/silverback-mono/compare/@-amazeelabs/silverback-drupal-tests@1.1.126...@-amazeelabs/silverback-drupal-tests@1.1.127) (2024-05-10)
 
 **Note:** Version bump only for package @-amazeelabs/silverback-drupal-tests